requirements.txt
playwright==1.43.0
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.1
googletrans==4.0.0-rc1
fastapi==0.110.0
uvicorn==0.29.0
jieba==0.42.1